+++ apps/frontend/homepage-app/app/login/mfa/mfa-setup.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/buttons/button';
 import Image from 'next/image';
 
 export interface MfaSetupProps {
