+++ apps/frontend/homepage-app/app/login/mfa/mfa-verify.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/buttons/button';
 
 export interface MfaVerifyProps {
   /**
