+++ apps/frontend/homepage-app/app/login/mfa/optional-mfa-prompt.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/buttons/button';
 
 export interface OptionalMfaPromptProps {
   /**
