+++ apps/frontend/homepage-app/app/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import CanvasBackground from '@/components/ui/backgrounds/canvasBackground';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/buttons/button';
 import Link from 'next/link';
 import useEmblaCarousel from 'embla-carousel-react';
 import { useEffect, useRef, useCallback } from 'react';
